Development

+++ app/models/groups/sql.go
@@ -15,23 +15,19 @@ func addGroup() string {
 	`
 }
 
+func modifyGroup() string {
+	return `
+		UPDATE Groups SET name=$3, description=$4
+		WHERE id=$1 AND owner_id=$2
+		RETURNING id
+	`
+}
+
 func addWord() string {
 	return `
 		INSERT INTO Words (owner_id, word, description)
 		VALUES ($1, $2, $3)
 		RETURNING id
-	`
-}
-
-func wordById() string {
-	return `
-		SELECT id, owner_id, word, description FROM Words WHERE id=$1;
-	`
-}
-
-func wordItemsByWordId() string {
-	return `
-		SELECT I.id, I.word, I.description FROM Words W, WordItems I Where W.id=$1 AND W.id=I.word_id;
 	`
 }
 
@@ -43,11 +39,20 @@ func addWordItem() string {
 	`
 }
 
+func deleteGroupLinks(ids g.GroupIds) string {
+	array := strings.Trim(strings.Join(strings.Fields(fmt.Sprint(ids.Ids)), ", "), "[]")
+	sql := fmt.Sprintf(`
+		DELETE FROM GroupLinks WHERE group_id IN (%s)
+	`, array)
+	return sql
+}
+
 func deleteGroups(ids g.GroupIds) string {
 	// I know, there is a risk for sql infjection attack, but this should be ok for int an array
 	array := strings.Trim(strings.Join(strings.Fields(fmt.Sprint(ids.Ids)), ", "), "[]")
 	sql := fmt.Sprintf(`
-		DELETE FROM Groups WHERE id IN (%s)
+		DELETE FROM Groups WHERE id IN (%s) AND owner_id=$1
+		RETURNING id
 	`, array)
 	return sql
 }
